requirements.txt
asgiref==3.4.1
Django==3.2.5
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djongo==1.3.6
dnspython==1.16.0
PyJWT==2.1.0
pymongo==3.12.0
pytz==2021.1
sqlparse==0.2.4
